cleanup: remove legacy shell analyzers and AST rollout flags
